+++ packages/camera/lib/camera.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
@@ -115,8 +116,27 @@ class CameraPreview extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    // ISSUE: The Texture seems to return a buffer in portrait in Android (width and height are inverted) and a buffer in landscape on iOS (which respects the aspect ratio).
+    // The ideal code (if the Texture was oriented properly in landscape for both Android and iOS) would be:
+    /*
+    RotatedBox(
+      quarterTurns: controller.description.sensorOrientation ~/ 90,
+      child: AspectRatio(
+        aspectRatio: controller.value.aspectRatio,
+        child: Texture(textureId: controller._textureId),
+      ),
+    )
+    */
     return controller.value.isInitialized
-        ? Texture(textureId: controller._textureId)
+        ? RotatedBox(
+            quarterTurns: Platform.isAndroid ? 0 : 1,
+            child: AspectRatio(
+              aspectRatio: Platform.isAndroid
+                  ? 1 / controller.value.aspectRatio
+                  : controller.value.aspectRatio,
+              child: Texture(textureId: controller._textureId),
+            ),
+          )
         : Container();
   }
 }
@@ -156,12 +176,14 @@ class CameraValue {
   /// The size of the preview in pixels.
   ///
   /// Is `null` until  [isInitialized] is `true`.
+  ///
+  /// The preview size may be smaller than the size of the recorded video (for performance issues). But the aspect ratio will be preserved.
   final Size previewSize;
 
-  /// Convenience getter for `previewSize.height / previewSize.width`.
+  /// Convenience getter for `previewSize.width / previewSize.height`.
   ///
   /// Can only be called when [initialize] is done.
-  double get aspectRatio => previewSize.height / previewSize.width;
+  double get aspectRatio => previewSize.width / previewSize.height;
 
   bool get hasError => errorDescription != null;